+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,8 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\Kalaha\\NeuralNetwork"
+  ],
+  "SelectedNode": "\\Kalaha.sln",
+  "PreviewInSolutionExplorer": false
+}
