+++ Kalaha/Kalaha.csproj
@@ -55,13 +55,23 @@
     <Compile Include="HumanPlayer.cs" />
     <Compile Include="IChooseMove.cs" />
     <Compile Include="KalahaBoard.cs" />
-    <Compile Include="KalahaBoardDisplay.cs" />
-    <Compile Include="KalahaBoardDisplay.Designer.cs">
-      <DependentUpon>KalahaBoardDisplay.cs</DependentUpon>
+    <Compile Include="KalahaBoardDisplay_Dani.cs">
+      <SubType>Form</SubType>
     </Compile>
-    <Compile Include="KalahaBoardDisplaySTD.cs" />
-    <Compile Include="KalahaBoardDisplaySTD.Designer.cs">
-      <DependentUpon>KalahaBoardDisplaySTD.cs</DependentUpon>
+    <Compile Include="KalahaBoardDisplay_Dani.Designer.cs">
+      <DependentUpon>KalahaBoardDisplay_Dani.cs</DependentUpon>
+    </Compile>
+    <Compile Include="KalahaBoardDisplaySTD_Dani.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="KalahaBoardDisplaySTD_Dani.Designer.cs">
+      <DependentUpon>KalahaBoardDisplaySTD_Dani.cs</DependentUpon>
+    </Compile>
+    <Compile Include="KalahaBoardDisplay_STD.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="KalahaBoardDisplay_STD.Designer.cs">
+      <DependentUpon>KalahaBoardDisplay_STD.cs</DependentUpon>
     </Compile>
     <Compile Include="KalahaMatch.cs" />
     <Compile Include="KIPlayer.cs" />
@@ -84,13 +94,15 @@
     <None Include="app.config" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="KalahaBoardDisplay.resx">
-      <DependentUpon>KalahaBoardDisplay.cs</DependentUpon>
+    <EmbeddedResource Include="KalahaBoardDisplay_Dani.resx">
+      <DependentUpon>KalahaBoardDisplay_Dani.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="KalahaBoardDisplaySTD.resx">
-      <DependentUpon>KalahaBoardDisplaySTD.cs</DependentUpon>
+    <EmbeddedResource Include="KalahaBoardDisplaySTD_Dani.resx">
+      <DependentUpon>KalahaBoardDisplaySTD_Dani.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="KalahaBoardDisplay_STD.resx">
+      <DependentUpon>KalahaBoardDisplay_STD.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
-  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
