+++ Kalaha/KalahaBoardDisplay_STD.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Windows.Forms;
+
+namespace Kalaha
+{
+    public partial class KalahaBoardDisplay_STD : Form
+    {
+        public KalahaBoardDisplay_STD(int StartValue)
+        {
+            InitializeComponent();
+        }
+
+        public void Refresh(int[] Slots)
+        {
+
+        }
+    }
+}
